pyproject.toml
[tool.poetry]
name = "ttc-python-etl"
version = "0.1.0"
description = ""
authors = ["Thomas Tunc <28352452+thomastunc@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
duckdb = "^1.1.1"
duckdb-engine = "^0.13.2"
python-dotenv = "^1.0.1"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
